+++ bull-bean-transformer/src/main/java/com/hotels/beans/transformer/AbstractTransformer.java
@@ -17,15 +17,12 @@
 package com.hotels.beans.transformer;
 
 import static java.util.Arrays.asList;
-import static java.util.Arrays.stream;
 
 import static com.hotels.transformer.cache.CacheManagerFactory.getCacheManager;
 import static com.hotels.transformer.validator.Validator.notNull;
 
 import java.util.Map;
-import java.util.function.Function;
 
-import com.hotels.beans.BeanUtils;
 import com.hotels.beans.conversion.analyzer.ConversionAnalyzer;
 import com.hotels.transformer.cache.CacheManager;
 import com.hotels.transformer.model.FieldMapping;
@@ -105,17 +102,6 @@ abstract class AbstractTransformer implements BeanTransformer {
     /**
      * {@inheritDoc}
      */
-    @Deprecated(since = "1.6.0", forRemoval = true)
-    public final Transformer withFieldMapping(final com.hotels.beans.model.FieldMapping... fieldMapping) {
-        com.hotels.transformer.model.FieldMapping[] fieldMappings = stream(fieldMapping)
-                .map(fm -> new com.hotels.transformer.model.FieldMapping(fm.getSourceFieldName(), fm.getDestFieldName()))
-                .toArray(com.hotels.transformer.model.FieldMapping[]::new);
-        return withFieldMapping(fieldMappings);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
     @Override
     public final void removeFieldMapping(final String destFieldName) {
         notNull(destFieldName, "The field name for which the mapping has to be removed cannot be null!");
@@ -140,17 +126,6 @@ abstract class AbstractTransformer implements BeanTransformer {
             fieldsTransformers.put(transformer.getDestFieldName(), transformer);
         }
         return this;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Deprecated(since = "1.6.0", forRemoval = true)
-    public final Transformer withFieldTransformer(final com.hotels.beans.model.FieldTransformer... fieldTransformer) {
-        Function<com.hotels.beans.model.FieldTransformer, FieldTransformer> transformerFunction = BeanUtils.getTransformer(FieldTransformer.class);
-        com.hotels.transformer.model.FieldTransformer[] fieldTransformers = stream(fieldTransformer)
-                .map(transformerFunction).toArray(com.hotels.transformer.model.FieldTransformer[]::new);
-        return withFieldTransformer(fieldTransformers);
     }
 
     /**