+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
-### [1.5.1] TBD
+### [1.6.0] TBD
+### Removed
+* Removed deprecated module `bean-utils-library`
+* The following deprecated classes has been removed:
+    * `com.hotels.beans.model.FieldMapping`
+    * `com.hotels.beans.model.FieldTransformer`
+    * `com.hotels.beans.Transformer`
+    
+### [1.5.1] 2019.09.02
 #### Changed
 * **The module `bean-utils-library` has been deprecated and will be no longer available since version `1.6.0`, use `bull-bean-transformer` instead.**
     ~~~